ml/requirements.txt
fastapi==0.110.0
uvicorn==0.27.1
python-multipart==0.0.9
pytest==8.1.1
pytest-mock==3.14.0
pillow==10.2.0
pytesseract==0.3.10
numpy==1.26.4
sentencepiece==0.2.0
protobuf==4.25.3
transformers==4.38.2
torch==2.2.1
redis==5.0.3
bitsandbytes==0.42.0
accelerate==0.28.0
nltk==3.8.1
keybert==0.8.3
psycopg2-binary==2.9.9
python-magic==0.4.27
pytest-asyncio==0.23.5 